+++ src/modules/products/components/ProductsContent.js
@@ -4,6 +4,7 @@ import get from 'lodash/get';
 
 import history from 'lib/history';
 import CustomTable from 'modules/shared/components/CustomTable';
+const screenWidth = window.innerWidth;
 
 const columns = [
   {
@@ -23,17 +24,34 @@ const columns = [
     field: 'weight',
     formatter: ({ packagingOptions }) => {
       const weight = get(packagingOptions, '0.grossWeightGrams', 0);
-      return `${weight} g`;
+      if (screenWidth < 750) {
+        return <div style={{ textAlign: 'left' }}> {weight} g </div>;
+      }
+      return <div style={{ textAlign: 'center' }}> {weight} g</div>;
     },
   },
   {
     title: 'Price',
     field: 'defaultPrice',
-    formatter: ({ defaultPrice = 0 }) => `€ ${defaultPrice.toFixed(2)}`,
+    formatter: ({ defaultPrice = 0 }) => {
+      if (screenWidth < 750) {
+        return (
+          <div style={{ textAlign: 'left' }}> {defaultPrice.toFixed(2)}€ </div>
+        );
+      }
+      return (
+        <div style={{ textAlign: 'right' }}> {defaultPrice.toFixed(2)}€</div>
+      );
+    },
   },
 ];
 
-const ProductsContent = ({ products, getData, isLoading, setSortByInCaller }) => {
+const ProductsContent = ({
+  products,
+  getData,
+  isLoading,
+  setSortByInCaller,
+}) => {
   const clickRow = ({ _id }) => {
     history.push(`/products/${_id}`);
   };