+++ src/modules/transactions/components/TransactionsContent.js
@@ -87,14 +87,14 @@ const columns = [
       if (price === 0) {
         return '';
       }
-      return <div style={{ textAlign: 'right' }}> {price}€ </div>;
+      return <div style={{ textAlign: 'right' }}> {price.toFixed(2)}€ </div>;
     },
   },
   {
     title: 'Total Price',
     field: 'total',
     formatter: ({ total }) => (
-      <div style={{ textAlign: 'right' }}> {total}€ </div>
+      <div style={{ textAlign: 'right' }}> {total.toFixed(2)}€ </div>
     ),
   },
   // {